+++ docs/app.js
@@ -38,15 +38,6 @@ new Vue({
             {label: '10 users', value: 10},
             {label: '12 users', value: 12},
         ],
-        roleOptions: [
-            {label: 'Administrators', value: 'admin'},
-            {label: 'Moderators', value: 'moderator'},
-            {label: 'Users', value: 'user'},
-        ],
-        statusOptions: [
-            {label: 'All', value: ''},
-            {label: 'Banned', value: 'banned'},
-        ],
     },
     components: {AsyncSearch},
 })