+++ src/components/Checkbox.vue
@@ -1,25 +1,13 @@
-<style>
-</style>
-
 <template>
-    <div>
-        <div class="as-checkbox-item" v-for="option in options">
-            <label class="as-checkbox-label">
-                <input type="checkbox" v-model="value" :value="getValueForOption(option)" class="as-checkbox-checkbox">
-
-                {{ getLabelForOption(option) }}
-            </label>
-        </div>
-    </div>
+    <input type="checkbox" v-model="theValue" :value="value" class="as-checkbox-checkbox">
 </template>
 
 <script>
     import ParamMixin from '../mixins/ParamMixin'
-    import OptionsMixin from '../mixins/OptionsMixin'
     import ValueMixin from '../mixins/ValueMixin'
 
     export default {
-        mixins: [ParamMixin, OptionsMixin, ValueMixin()],
+        mixins: [ParamMixin, ValueMixin('theValue')],
         props: {
             defaultValue: {
                 type: Array,
@@ -27,6 +15,24 @@
                     return []
                 },
             },
+            value: {
+                type: String,
+                required: true,
+            },
+            checked: {
+                type: Boolean,
+                default: false,
+            },
+        },
+        created() {
+            if (!this.checked) {
+                return
+            }
+            let defaultValue = this.searchStore.getDefaultValue(this.name)
+
+            defaultValue.push(this.value)
+
+            this.searchStore.setDefaultValue(this.name, defaultValue)
         },
     }
 </script>