+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.2.0
 - Support for object responses
+- The `Search` component no longer requires an explicit wrapper node for its children
 
 ## 1.1.0
 - The fields used for the label and value of Checkbox/Radio/Select components